requirements.txt
flask 
flask-sqlalchemy
Flask-Migrate
python-dotenv
flask-bcrypt
flask-jwt-extended